sched/rt: Refactor rt_energy_aware_wake_cpu()
Refactor the RT task placement code similar to fair task
placement. The CPUs are iterated from the minimum capacity
CPU's sched group in the highest sched domain which has
energy cost data. If we find an eligible in the lower
capacity group, skip iterating over the other groups.

There is no functionality change with this patch. The
refactoring is done to prepare the code ready for
implementing placement boost support.

Change-Id: I66dd276232f38554febbdf905b31b6656cad98ff
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: Yousef Algadri <yusufgadrie@gmail.com>
Signed-off-by: DennySPB <dennyspb@gmail.com>
Signed-off-by: prorooter007 <shreyashwasnik112@gmail.com>