msm: kgsl: Implement a list of imported memory
With ION now using the dma attachment's priv pointer, we need another
way to associate the imported buffers with the mem entries in kgsl.
Create a list of all dma buffers imported into kgsl, identified by
the first page in the scatterlist. Use this list to determine how
many times a given buffer was mapped as an egl surface and as an egl
image.

Change-Id: Ic1d1d89de8fb5ba7264825810515efd03215feca
Signed-off-by: Harshdeep Dhatt <hdhatt@codeaurora.org>
Signed-off-by: Lynus Vaz <lvaz@codeaurora.org>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: Dusan Uveric <dusan.uveric9@gmail.com>
Signed-off-by: prorooter007 <shreyashwasnik112@gmail.com>