f2fs/rapid_gc: Do not wait until all invalid blocks are discarded
Undiscarded blocks may not reach 0 for various reasons, such as
discard granularity settings and discard thread terminating before GC ends,
and when that happens, GC thread enters an indefinite loop.

Remove this logic entirely.

This is not much of an issue since discard thread does not rely on rapid GC.

Change-Id: Ida35007f0913f00f72a81537e47313d56c1887a6
Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>
Signed-off-by: prorooter007 <shreyashwasnik112@gmail.com>